board: otter: ottercast-amp: fix incorrect pll-audio postdivider
One of the previous changes moved the audio-pll to SDM, In the process
the psotdivider was supposed to be configutred to be one. However, that
did not happen and caused the audio-pll clock to run at a quater of the
inteded speed. Fix that by forcing the postdivider to oen.