board: otter: disable DNSSEC in systemd-resolved
This commit disables DNSSEC validation in systemd-resolved,
because the ottercasts have no HW real time clock. After a reboot
the time will be wrong, DNSSEC signatures for the timeservers will be
expired/not yet valid and the hostnames for the timeservers can not
be resolved. This leads to an infinite loop of not working NTP,
not working DNSSEC/DNS, which will break any application software.