package: add support for AP6255 dual-band WiFi module
The default configuration for both the OtterCast AMP and OtterCast Audio V2
includes a single band 2.4GHz AP6236 WiFi module only.
Since 2.4GHz WiFi is not that great for streaming in densly populated areas
this commit adds support for the AP6255 dual-band WiFi module, which allows
the OtterCast to work a lot more smothly, using the 5GHz WiFI band.

Fixes #20